Added experimental JEID support

+++ main/java/jaredbgreat/climaticbiome/generation/biome/IBiomeSpecifier.java
@@ -3,7 +3,7 @@ package jaredbgreat.climaticbiome.generation.biome;
 import jaredbgreat.climaticbiome.generation.generator.ChunkTile;
 
 public interface IBiomeSpecifier {	
-	public int getBiome(ChunkTile tile);
+	public long getBiome(ChunkTile tile);
 	public boolean isEmpty();
 
 }
