+++ main/java/jaredbgreat/climaticbiome/generation/biome/biomes/GetColdPlains.java
@@ -23,7 +23,7 @@ public class GetColdPlains implements IBiomeSpecifier {
 	}
 
 	@Override
-	public int getBiome(ChunkTile tile) {
+	public long getBiome(ChunkTile tile) {
 		if((tile.getBiomeSeed() % 3) == 0) {
 			return 13;  // TODO: Once GetAlpine is more generic use that
 		}
