+++ main/java/jaredbgreat/climaticbiome/generation/biome/biomes/GetJungle.java
@@ -32,7 +32,7 @@ public class GetJungle implements IBiomeSpecifier {
 
 
 	@Override
-	public int getBiome(ChunkTile tile) {
+	public long getBiome(ChunkTile tile) {
 		int role1 = tile.getBiomeSeed() % 5;
 		int role2 = tile.getBiomeSeed() % 11;
 		int role3 = tile.getBiomeSeed() % 7;
